+++ Pod/Classes/Configurations/VSLAccountConfiguration.m
@@ -14,6 +14,13 @@
         self.dropCallOnRegistrationFailure = NO;
         self.mediaStunType = VSLStunUseRetryOnFailure;
         self.sipStunType = VSLStunUseDefault;
+        self.contactRewriteMethod = VSLContactRewriteMethodAlwaysUpdate;
+        self.ipAddressChangeShutdownTransport = YES;
+        self.ipAddressChangeHangupAllCalls = NO;
+        self.ipAddressChangeReinviteFlags = VSLReinviteFlagsReinitMedia | VSLReinviteFlagsUpdateVia | VSLReinviteFlagsUpdateContact;
+        self.contactUseSrcPort = YES;
+        self.allowViaRewrite = YES;
+        self.allowContactRewrite = YES;
     }
     return self;
 }